Move PermissionUtil to internal namespace, remove test case
Windsor doesn't have internals visible to test projects, and not worth adding for one test case. This is code brought across unchanged from `Castle.Core` and is likely to be deprecated soon, as CAS is phased out.

#612 - Updating Windsor to support Castle.Core@5.0.0 and modern TFMs
